refactor: accordion.js
reviewed: Emmanuel Oliveira
