+++ scripts/index.js
@@ -1,7 +1,6 @@
 import scrollSoft from "./modules/initScroll.js"
-import initAccordion from "./modules/accordion.js"
+import Accordion from "./modules/accordion.js"
 import activeTabs from "./modules/activeTabs.js"
-import initAnimationScroll from "./modules/animationScroll.js"
 import initModal from "./modules/modal.js"
 import InitTolltip from "./modules/tooltip.js"
 import initDropdownMenu from "./modules/dropdown-menu.js"
@@ -9,12 +8,14 @@ import initMenuMobile from "./modules/menu-mobile.js"
 import initOperation from "./modules/operation.js"
 import initFetchAnimais from "./modules/fetch-animais.js"
 import initFetchBitcoin from "./modules/fetch-bitcoin.js"
+import initAnimationScroll from "./modules/animationScroll.js"
 
 const ScrollSoft = new scrollSoft('[data-menu="smoth"] a[href^="#"]')
 ScrollSoft.init()
+const accordion = new Accordion("[data-anime='accordion'] dt")
+accordion.init()
 
 activeTabs()
-initAccordion()
 initAnimationScroll()
 initModal()
 InitTolltip()
