+++ scripts/modules/accordion.js
@@ -1,17 +1,25 @@
-export default function initAccordion() {
-  const accordionList = document.querySelectorAll("[data-anime='accordion'] dt")
-  const activeClass = "active"
-
-  function activeAccordion() {
-    this.classList.toggle(activeClass)
-    this.nextElementSibling.classList.toggle(activeClass)
+export default class Accordion {
+  constructor(list) {
+    this.accordionList = document.querySelectorAll(list)
+    this.activeClass = "active"
   }
 
-  if (accordionList.length) {
-    accordionList[0].classList.add(activeClass)
-    accordionList[0].nextElementSibling.classList.add(activeClass)
-    accordionList.forEach((item) => {
-      item.addEventListener("click", activeAccordion)
+  toggleAccordion(item) {
+    item.classList.toggle(this.activeClass)
+    item.nextElementSibling.classList.toggle(this.activeClass)
+  }
+  //Adiciona os Eventos ao Accordion
+  addAccordionEvent() {
+    this.accordionList.forEach((item) => {
+      item.addEventListener("click", () => this.toggleAccordion(item))
     })
+  }
+  // Iniciar função
+  init() {
+    if (this.accordionList.length) {
+      //Ativar primeiro Item
+      this.toggleAccordion(this.accordionList[0])
+      this.addAccordionEvent()
+    }
   }
 }
