+++ scripts/modules/backup/accordion-bkp.js
@@ -1,0 +1,17 @@
+export default function initAccordion() {
+  const accordionList = document.querySelectorAll("[data-anime='accordion'] dt")
+  const activeClass = "active"
+
+  function activeAccordion() {
+    this.classList.toggle(activeClass)
+    this.nextElementSibling.classList.toggle(activeClass)
+  }
+
+  if (accordionList.length) {
+    accordionList[0].classList.add(activeClass)
+    accordionList[0].nextElementSibling.classList.add(activeClass)
+    accordionList.forEach((item) => {
+      item.addEventListener("click", activeAccordion)
+    })
+  }
+}
